minimal refactor of collections pane

+++ client/components/DashboardCollections/CollectionRow.js
@@ -7,6 +7,7 @@ import communityType from 'types/community';
 import { getSchemaForKind } from 'shared/collections/schemas';
 
 import LinkedPageSelect from './LinkedPageSelect';
+import ConfirmDialog from '../ConfirmDialog/ConfirmDialog';
 
 const propTypes = {
 	collection: collectionType.isRequired,
@@ -17,6 +18,7 @@ const propTypes = {
 
 const CollectionRow = ({ communityData, collection, onUpdateCollection, onDeleteCollection }) => {
 	const schema = getSchemaForKind(collection.kind);
+	const label = schema.label.singular;
 	const canEditMetadata = schema.metadata.length > 0;
 	return (
 		<div key={`collection-${collection.id}`} className="collection-wrapper">
@@ -64,13 +66,21 @@ const CollectionRow = ({ communityData, collection, onUpdateCollection, onDelete
 			>
 				Private
 			</Checkbox>
-			<button
-				type="button"
-				className="bp3-button bp3-icon-small-cross bp3-minimal"
-				onClick={() => {
+			<ConfirmDialog
+				text={`Are you sure you want to delete this ${label}?`}
+				confirmLabel="Delete"
+				onConfirm={() => {
 					onDeleteCollection(collection.id);
 				}}
-			/>
+			>
+				{({ open }) => (
+					<button
+						type="button"
+						className="bp3-button bp3-icon-small-cross bp3-minimal"
+						onClick={open}
+					/>
+				)}
+			</ConfirmDialog>
 		</div>
 	);
 };
